+++ changelog.md
@@ -8,7 +8,7 @@
 * Upgraded mongodb drivers from 4.10.2 to 4.11.1
 * Upgraded lettuce core from 6.2.6.RELEASE to 6.3.0.RELEASE
 * Upgraded spring-aspects from 6.0.10 to 6.1.1
-* Upgraded jackson from 2.15.2 to 2.16.0
+* Upgraded jackson from 2.15.2 to 2.15.3
 
 ### 0.16.11 (2023-12-01)
 * Removed `isFinalError` method from `ErrorInfo` used by `RetryStrategy`, use `isLastAttempt()` instead.