+++ changelog.md
@@ -1,6 +1,7 @@
 ### Changelog next version
 * spring-boot-starter-mongodb no longer autoconfigures itself by just importing the library in the classpath, instead you need to bootstrap by annotating your Spring Boot class with @EnableOccurrent.   
-* Fixed bug in spring-boot-starter-mongodb module in which it didn't automatically configured MongoDB. 
+* Fixed bug in spring-boot-starter-mongodb module in which it didn't automatically configured MongoDB.
+* Domain event subscriptions now accepts metadata as the first parameter, besides just the event. The metadata currently contains the stream version and stream id, which can be useful when building projections. 
 * Upgraded from Kotlin 1.9.20 to 1.9.22
 * Upgraded amqp-client from 5.16.0 to 5.20.0
 * Upgraded Spring Boot from 3.1.4 to 3.2.1